feat: [FN-258] 학습 모드 설정 컴포넌트 분리
