+++ src/features/setting-study-mode/ui/study-settings.tsx
@@ -3,11 +3,8 @@ import {
   ButtonCheckbox,
   ButtonCheckboxGroupField,
 } from "@/shared/components/button-checkbox";
-import { Checkbox } from "@/shared/components/checkbox";
-import { Description, FormTitle } from "@/shared/components/form";
-import { NumberInput } from "@/shared/components/number-input";
+import { FormTitle } from "@/shared/components/form";
 import { Label } from "@/shared/components/label";
-import { ToggleGroup } from "@/shared/components/toggle-group";
 import { useNavigate } from "@tanstack/react-router";
 import { BookOpenCheck, Brain, RotateCcw, Settings2Icon } from "lucide-react";
 import { useController, useForm, useWatch } from "react-hook-form";
@@ -17,6 +14,8 @@ import {
   type StudySettingsFormField,
 } from "../model/form.schema";
 import { useEffect } from "react";
+import { MemorizeSettingsForm } from "./memorize-settings-form";
+import { TestSettingsForm } from "./test-settings-form";
 
 type StudySettingsProps = {
   groupId: number;
@@ -66,26 +65,7 @@ const StudySettings = ({
     control,
   });
 
-  const { field: orderTypeField } = useController({
-    name: "orderType",
-    control,
-  });
-
-  const { field: navigationTypeField } = useController({
-    name: "navigationType",
-    control,
-  });
-
-  const { field: testModeField } = useController({
-    name: "testMode",
-    control,
-  });
-
   const mode = useWatch({ control, name: "mode" });
-  const navigationType = useWatch({ control, name: "navigationType" });
-  const isUnlimitedRepeat = useWatch({ control, name: "isUnlimitedRepeat" });
-  const isUnlimitedTime = useWatch({ control, name: "isUnlimitedTime" });
-  const testMode = useWatch({ control, name: "testMode" });
 
   // 모드 변경 시 기본값 설정
   useEffect(() => {
@@ -161,221 +141,21 @@ const StudySettings = ({
 
           {/* 암기 모드 설정 */}
           {mode === "memorize" && (
-            <>
-              {/* 반복 횟수 */}
-              <fieldset className="space-y-3">
-                <Label className="text-sm font-medium">반복 횟수</Label>
-                <div className="flex items-center gap-2">
-                  <NumberInput
-                    min={1}
-                    max={100}
-                    className="w-24"
-                    disabled={isUnlimitedRepeat}
-                    {...register("repeatCount", {
-                      valueAsNumber: true,
-                    })}
-                  />
-                  <span className="text-sm">회</span>
-                </div>
-                <div className="flex items-center gap-2">
-                  <Checkbox
-                    id="unlimited-repeat"
-                    {...register("isUnlimitedRepeat")}
-                  />
-                  <label
-                    htmlFor="unlimited-repeat"
-                    className="text-sm cursor-pointer"
-                  >
-                    제한 없음
-                  </label>
-                </div>
-                {"repeatCount" in errors && errors.repeatCount && (
-                  <p className="text-sm text-red-500 mt-1">
-                    {errors.repeatCount.message}
-                  </p>
-                )}
-              </fieldset>
-
-              {/* 카드 순서와 페이지 넘김 */}
-              <div className="w-full flex justify-stretch items-stretch gap-2">
-                <fieldset className="space-y-3 grow">
-                  <Label className="text-sm font-medium">카드 순서</Label>
-                  <ToggleGroup
-                    name="orderType"
-                    value={orderTypeField.value}
-                    onChange={orderTypeField.onChange}
-                    onBlur={orderTypeField.onBlur}
-                    options={[
-                      { value: "sequential", label: "순차 진행" },
-                      { value: "random", label: "랜덤 섞기" },
-                    ]}
-                  />
-                  {errors.orderType && (
-                    <p className="text-sm text-red-500 mt-1">
-                      {errors.orderType.message}
-                    </p>
-                  )}
-                </fieldset>
-
-                <fieldset className="space-y-3 grow">
-                  <Label className="text-sm font-medium">페이지 넘김</Label>
-                  <ToggleGroup
-                    name="navigationType"
-                    value={navigationTypeField.value}
-                    onChange={navigationTypeField.onChange}
-                    onBlur={navigationTypeField.onBlur}
-                    options={[
-                      { value: "manual", label: "수동" },
-                      { value: "auto", label: "자동" },
-                    ]}
-                  />
-                  {"navigationType" in errors && errors.navigationType && (
-                    <p className="text-sm text-red-500 mt-1">
-                      {errors.navigationType.message}
-                    </p>
-                  )}
-                </fieldset>
-              </div>
-
-              {/* 자동 넘김 시간 */}
-              {navigationType === "auto" && (
-                <div className="space-y-3 bg-blue-50 p-5 rounded-lg">
-                  <Label
-                    htmlFor="autoTimerSeconds"
-                    className="text-sm font-medium"
-                  >
-                    자동 넘김 시간
-                  </Label>
-                  <Description>
-                    카드가 자동으로 넘어가는 시간을 설정해주세요
-                  </Description>
-                  <div className="flex items-center gap-2">
-                    <NumberInput
-                      id="autoTimerSeconds"
-                      className="w-24"
-                      {...register("autoTimerSeconds", {
-                        valueAsNumber: true,
-                      })}
-                    />
-                    <span className="text-sm">초</span>
-                  </div>
-                  {"autoTimerSeconds" in errors && errors.autoTimerSeconds && (
-                    <p className="text-sm text-red-500 mt-1">
-                      {errors.autoTimerSeconds.message}
-                    </p>
-                  )}
-                </div>
-              )}
-            </>
+            <MemorizeSettingsForm
+              control={control}
+              register={register}
+              errors={errors}
+            />
           )}
 
           {/* 시험 모드 설정 */}
           {mode === "test" && (
-            <>
-              {/* 시험 시간 */}
-              <fieldset className="space-y-3">
-                <Label className="text-sm font-medium">시험 시간</Label>
-                <div className="flex items-center gap-2">
-                  <NumberInput
-                    min={1}
-                    max={180}
-                    className="w-24"
-                    disabled={isUnlimitedTime}
-                    {...register("testTimeMinutes", {
-                      valueAsNumber: true,
-                    })}
-                  />
-                  <span className="text-sm">분</span>
-                </div>
-                <div className="flex items-center gap-2">
-                  <Checkbox
-                    id="unlimited-time"
-                    {...register("isUnlimitedTime")}
-                  />
-                  <label
-                    htmlFor="unlimited-time"
-                    className="text-sm cursor-pointer"
-                  >
-                    제한 없음
-                  </label>
-                </div>
-                {"testTimeMinutes" in errors && errors.testTimeMinutes && (
-                  <p className="text-sm text-red-500 mt-1">
-                    {errors.testTimeMinutes.message}
-                  </p>
-                )}
-              </fieldset>
-
-              {/* 시험 순서와 시험 모드 */}
-              <div className="w-full flex justify-stretch items-start gap-2">
-                <fieldset className="space-y-3 grow">
-                  <Label className="text-sm font-medium">시험 순서</Label>
-                  <ToggleGroup
-                    name="orderType"
-                    value={orderTypeField.value}
-                    onChange={orderTypeField.onChange}
-                    onBlur={orderTypeField.onBlur}
-                    options={[
-                      { value: "sequential", label: "순차 시험" },
-                      { value: "random", label: "랜덤 시험" },
-                    ]}
-                  />
-                  {errors.orderType && (
-                    <p className="text-sm text-red-500 mt-1">
-                      {errors.orderType.message}
-                    </p>
-                  )}
-                </fieldset>
-
-                <fieldset className="space-y-3 grow">
-                  <Label className="text-sm font-medium">시험 모드</Label>
-                  <ToggleGroup
-                    name="testMode"
-                    value={testModeField.value}
-                    onChange={testModeField.onChange}
-                    onBlur={testModeField.onBlur}
-                    options={[
-                      { value: "all", label: "전체 시험" },
-                      { value: "random", label: "랜덤 뽑기" },
-                    ]}
-                  />
-                  {"testMode" in errors && errors.testMode && (
-                    <p className="text-sm text-red-500 mt-1">
-                      {errors.testMode.message}
-                    </p>
-                  )}
-
-                  {/* 랜덤 뽑기 개수 */}
-                  {testMode === "random" && (
-                    <div className="space-y-3 mt-3">
-                      <Label htmlFor="randomPickCount" className="text-sm font-medium">
-                        랜덤 뽑기 개수
-                      </Label>
-                      <Description>
-                        전체 {totalCardCount}개 중 몇 개를 시험 볼까요?
-                      </Description>
-                      <div className="flex items-center gap-2">
-                        <NumberInput
-                          id="randomPickCount"
-                          min={1}
-                          max={totalCardCount}
-                          className="w-24"
-                          {...register("randomPickCount", {
-                            valueAsNumber: true,
-                          })}
-                        />
-                        <span className="text-sm">개</span>
-                      </div>
-                      {"randomPickCount" in errors && errors.randomPickCount && (
-                        <p className="text-sm text-red-500 mt-1">
-                          {errors.randomPickCount.message}
-                        </p>
-                      )}
-                    </div>
-                  )}
-                </fieldset>
-              </div>
-            </>
+            <TestSettingsForm
+              control={control}
+              register={register}
+              errors={errors}
+              totalCardCount={totalCardCount}
+            />
           )}
         </div>
